JWT based auth filters implementation + versioning of Rest Endpoints

+++ src/main/java/com/springboot/blog/springbootblogrestapi/security/JwtTokenProvider.java
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
+
 import java.util.Date;
 
 @Component
@@ -15,7 +17,7 @@ public class JwtTokenProvider {
     private String jwtSecret;
 
     @Value("${app.jwt-expiration-milliseconds}")
-    private int jwtExpiration;
+    private long jwtExpiration;
 
     //generate Token
